apps/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.11
datefinder==0.7.3
et_xmlfile==2.0.0
fastapi==0.115.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.16.5
httpx==0.28.1
google-auth==2.37.0
supabase==2.19.0
tenacity==9.1.2
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
h11==0.14.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
macholib==1.16.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pefile==2023.2.7
pillow==11.0.0
preshed==3.0.9
pycparser==2.22
pydantic-settings
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller-hooks-contrib==2024.11
pyinstaller==6.11.1
PyMuPDF==1.25.1
pyparsing==3.2.0
PyPDF2==3.0.1
pypdf==5.1.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy==3.8.4
sqlparse==0.5.3
srsly==2.5.1
starlette==0.41.3
sympy==1.13.1
thinc==8.3.4
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
email-validator==2.2.0
pytest==8.4.2
pytest-asyncio==0.23.8
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
XlsxWriter==3.2.0
